refactor(tauri): centralize app wiring and enforce lint
