+++ src-tauri/src/lib.rs
@@ -1,4 +1,9 @@
+mod app;
 mod commands;
+mod shortcuts;
 pub mod state;
+mod tray;
+mod windows;
 
+pub use app::run;
 pub use commands::*;
