+++ src-tauri/src/state.rs
@@ -22,6 +22,12 @@ pub struct AppState {
     pub last_saved_note: Mutex<Option<LastSavedNote>>,
 }
 
+impl Default for AppState {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
 impl AppState {
     pub fn new() -> Self {
         Self {
